Three columns for app.html
Realised that we need mechanisms to select and clear coordinates for source and destination location. This was not in the original design, and as such I had to make some modifications to the basic layout, to give 3 columns in the date and time section. Also started implementing some buttons and selectors.