Uploading progress made
Various changes made. Now have three columns at the top of the app to hold:
1) Source
2) Destination
3) Output

Also have a field to show messages to the user, which is useful for debugging beyond the console.

Have a test image for the source pin, and can add it on to the map, but having issues removing it from the map on the clear coordinates function.