Application - BUG - Clear Buttons
Fixed the clear button issue, it was caused by the presence of 'global' before some variables. This originated from testing, so has likely broken some things to do with testing, however time and patience with this project are now expired, and need to go with what I have got. More testing and code refinement was a known issue before this bug came to light.